Correlation analysis of genes expression in each sample
Pearson correlation analysis was performed for each sample in the data set to asses the gene coverage and correlation.

The correlation matrix heatmap shows 5 distinct clusters with higher correlation coeficient values. There clusters correspond to different cancer types and, as expected, have higher correlation among themselves.

Additionally, it can be seen that all the samples have a >0.6 correlation value, which is a good sign. This shows that samples have been processed similarly and there is no strong technical noise that might bias the data